third_party/blink/web_tests/platform/linux-chrome/external/wpt/fetch/stale-while-revalidate/stale-script-expected.txt

This is a testharness.js-based test.
[FAIL] Cache returns stale resource
  assert_equals: (revalidated version loaded) expected 2 but got 1
Harness: the test ran to completion.
